Coding/Agentic AI/PhiData/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile
#
annotated-types==0.7.0
    # via pydantic
anyio==4.7.0
    # via
    #   groq
    #   httpx
    #   openai
    #   starlette
arxiv==2.1.3
    # via -r requirements.in
beautifulsoup4==4.12.3
    # via yfinance
certifi==2024.12.14
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.1
    # via requests
click==8.1.8
    # via
    #   duckduckgo-search
    #   typer
    #   uvicorn
colorama==0.4.6
    # via
    #   click
    #   tqdm
distro==1.9.0
    # via
    #   groq
    #   openai
duckduckgo-search==7.1.1
    # via -r requirements.in
fastapi==0.115.6
    # via -r requirements.in
feedparser==6.0.11
    # via arxiv
frozendict==2.4.6
    # via yfinance
gitdb==4.0.11
    # via gitpython
gitpython==3.1.43
    # via phidata
groq==0.13.1
    # via -r requirements.in
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
html5lib==1.1
    # via yfinance
httpcore==1.0.7
    # via httpx
httpx==0.28.1
    # via
    #   groq
    #   openai
    #   phidata
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
jiter==0.8.2
    # via openai
lxml==5.3.0
    # via
    #   duckduckgo-search
    #   yfinance
markdown-it-py==3.0.0
    # via rich
mdurl==0.1.2
    # via markdown-it-py
multitasking==0.0.11
    # via yfinance
numpy==2.2.1
    # via
    #   pandas
    #   yfinance
openai==1.58.1
    # via -r requirements.in
packaging==24.2
    # via -r requirements.in
pandas==2.2.3
    # via yfinance
peewee==3.17.8
    # via yfinance
phidata==2.7.6
    # via -r requirements.in
platformdirs==4.3.6
    # via yfinance
primp==0.9.2
    # via duckduckgo-search
pydantic==2.10.4
    # via
    #   fastapi
    #   groq
    #   openai
    #   phidata
    #   pydantic-settings
pydantic-core==2.27.2
    # via pydantic
pydantic-settings==2.7.0
    # via phidata
pygments==2.18.0
    # via rich
pypdf==5.1.0
    # via -r requirements.in
python-dateutil==2.9.0.post0
    # via pandas
python-dotenv==1.0.1
    # via
    #   -r requirements.in
    #   phidata
    #   pydantic-settings
pytz==2024.2
    # via
    #   pandas
    #   yfinance
pyyaml==6.0.2
    # via phidata
requests==2.32.3
    # via
    #   arxiv
    #   yfinance
rich==13.9.4
    # via
    #   phidata
    #   typer
sgmllib3k==1.0.0
    # via feedparser
shellingham==1.5.4
    # via typer
six==1.17.0
    # via
    #   html5lib
    #   python-dateutil
smmap==5.0.1
    # via gitdb
sniffio==1.3.1
    # via
    #   anyio
    #   groq
    #   openai
soupsieve==2.6
    # via beautifulsoup4
starlette==0.41.3
    # via fastapi
tomli==2.2.1
    # via phidata
tqdm==4.67.1
    # via openai
typer==0.15.1
    # via phidata
typing-extensions==4.12.2
    # via
    #   anyio
    #   fastapi
    #   groq
    #   openai
    #   phidata
    #   pydantic
    #   pydantic-core
    #   typer
tzdata==2024.2
    # via pandas
urllib3==2.3.0
    # via requests
uvicorn==0.34.0
    # via -r requirements.in
webencodings==0.5.1
    # via html5lib
yfinance==0.2.51
    # via -r requirements.in
